requirements.txt
asgiref==3.2.0
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
Django==2.1.2
django-allauth==0.39.1
django-rest-auth==0.9.5
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
oauthlib==3.0.2
psycopg2==2.8.3
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.6
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
Twisted==19.2.1
txaio==18.8.1
urllib3==1.25.3
zope.interface==4.6.0
